==> Code/R/secondTutorial/7.2.R <==
getwd();
workingDir = "C:/Users/Taghi/eclipseWorkspace/School/Project/Bio-Informatics-Internship/Code/R/secondTutorial";
setwd(workingDir);
library(WGCNA);
library(cluster);
options(stringsAsFactors = FALSE);

load("Simulated-RelatingToExt.RData");

ADJ1 = abs(cor(datExpr,use = "p"))^6
Alldegrees1 = intramodularConnectivity(ADJ1,colorh1)

head(Alldegrees1)

colorLevels = unique(colorh1)
sizeGrWindow(9,6)
par(mfrow = c(2,as.integer(0.5 + length(colorLevels)/2)))
par(mar = c(4,5,3,1))
for (i in c(1:length(colorLevels)))
{
	whichmodule = colorLevels[[i]];
	restrict1 = (colorh1 == whichmodule);
	verboseScatterplot(Alldegrees1$kWithin[restrict1],
			GeneSignificance[restrict1], col = colorh1[restrict1],
			main = whichmodule,
			xlab = "Connectivity", ylab = "Gene Significance", abline = TRUE)
}

# Overall connectivity of all genes to modules not just intramodular connectiviy
datKME = signedKME(datExpr, datME, outputColumnName = "MM.")

head(datKME)

FilterGenes = abs(GS1) > .2 & abs(datKME$MM.brown) > .8
table(FilterGenes)

dimnames(data.frame(datExpr))[[2]][FilterGenes]

sizeGrWindow(8,6)
par(mfrow = c(2,2))

which.color = "turquoise";
restrictGenes = colorh1==which.color
verboseScatterplot(Alldegrees1$kWithin[restrictGenes],
			(datKME[restrictGenes, paste("MM.",which.color,sep = "")])^6,
			col = which.color, xlab = "Intramodular Connectivity",
			ylab = "(Module Membership)^6")

which.color = "blue";
restrictGenes = colorh1==which.color
verboseScatterplot(Alldegrees1$kWithin[restrictGenes],
			(datKME[restrictGenes, paste("MM.",which.color,sep = "")])^6,
			col = which.color, xlab = "Intramodular Connectivity",
			ylab = "(Module Membership)^6")

which.color = "brown";
restrictGenes = colorh1==which.color
verboseScatterplot(Alldegrees1$kWithin[restrictGenes],
			(datKME[restrictGenes, paste("MM.",which.color,sep = "")])^6,
			col = which.color, xlab = "Intramodular Connectivity",
			ylab = "(Module Membership)^6")

which.color = "green";
restrictGenes = colorh1==which.color
verboseScatterplot(Alldegrees1$kWithin[restrictGenes],
			(datKME[restrictGenes, paste("MM.",which.color,sep = "")])^6,
			col = which.color, xlab = "Intramodular Connectivity",
			ylab = "(Module Membership)^6")

NS1=networkScreening(y=y, datME=datME, datExpr=datExpr,
oddPower=3, blockSize=1000, minimumSampleSize=4,
addMEy=TRUE, removeDiag=FALSE, weightESy=0.5)

# network screening analysis
mean(NoiseGeneIndicator[rank(NS1$p.Weighted,ties.method="first")<=100])
# standard analysis based on the correlation p-values (or Student T test)
mean(NoiseGeneIndicator[rank(NS1$p.Standard,ties.method="first")<=100])


topNumbers=c(10,20,50,100)
for (i in c(1:length(topNumbers)) )
{
print(paste("Proportion of noise genes in the top", topNumbers[i], "list"))
WGCNApropNoise=mean(NoiseGeneIndicator[rank(NS1$p.Weighted,ties.method="first")<=topNumbers[i]])
StandardpropNoise=mean(NoiseGeneIndicator[rank(NS1$p.Standard,ties.method="first")<=topNumbers[i]])
print(paste("WGCNA, proportion of noise=", WGCNApropNoise,
", Standard, prop. noise=", StandardpropNoise))
if (WGCNApropNoise< StandardpropNoise) print("WGCNA wins")
if (WGCNApropNoise==StandardpropNoise) print("both methods tie")
if (WGCNApropNoise>StandardpropNoise) print("standard screening wins")
}

rm(dissTOM); collectGarbage();

CorPrediction1 = data.frame(GS1,NS1$cor.Weighted)
cor.Weighted = NS1$cor.Weighted

sizeGrWindow(8,6)
verboseScatterplot(cor.Weighted,GS1, main = "Network-based weighted correlation versus Pearson correlation\n",
			col = truemodule, cex.main = 1.2)
abline(0,1)



set.seed(2)
nSamples2 = 2000
MEgreen = rnorm(nSamples2)
scaledy2 = MEgreen * ESgreen + sqrt(1-ESgreen^2)*rnorm(nSamples2)
y2 = ifelse(scaledy2 > median(scaledy2),2,1)
MEturquoise = ESturquoise * scaledy2 + sqrt(1-ESturquoise^2)*rnorm(nSamples2)
MEblue = .6 * MEturquoise + sqrt(1-.6^2) * rnorm(nSamples2)
MEbrown = ESbrown * scaledy2 + sqrt(1-ESbrown^2)*rnorm(nSamples2)
MEyellow = ESyellow * scaledy2 + sqrt(1-ESyellow^2)*rnorm(nSamples2)

ModuleEigengeneNetwork2 = data.frame(y=y2,MEturquoise,MEblue,MEbrown,MEgreen,MEyellow)

dat2 = simulateDatExpr5Modules(MEturquoise = ModuleEigengeneNetwork2$MEturquoise,MEblue = ModuleEigengeneNetwork2$MEblue,
				MEbrown = ModuleEigengeneNetwork2$MEbrown,MEyellow = ModuleEigengeneNetwork2$MEyellow,
				MEgreen = ModuleEigengeneNetwork2$MEgreen, simulateProportions = simulateProportions1,
				nGenes = nGenes1)

GS1 = as.numeric(cor(y,datExpr,use = "p"))

GS2 = as.numeric(cor(ModuleEigengeneNetwork2$y, dat2$datExpr,use = "p"))

sizeGrWindow(8,6)
par(mfrow = c(1,1))
verboseScatterplot(GS1,GS2,
		main = "Trait-based gene significance in test set vs. training set\n",
		xlab = "Training set gene significance", ylab = "Test set gene significance",
		col = truemodule, cex.main = 1.4)
